wip better file structure + task actions/reducer
