+++ screens/StartupScreen/StartupScreen.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+const StartupScreen = () => {
+    return (
+        <View>
+            <Text>Startup Screen</Text>
+        </View>
+    );
+};
+
+export default StartupScreen;
