+++ screens/StartupScreen/index.js
@@ -1,12 +1,1 @@
-import React from "react";
-import { View, Text } from "react-native";
-
-const StartupScreen = () => {
-    return (
-        <View>
-            <Text>Startup Screen</Text>
-        </View>
-    );
-};
-
-export default StartupScreen;
+export { default } from "./StartupScreen";
