+++ screens/TimerScreen/ControlBar/ControlBar.js
