+++ screens/TimerScreen/Timer/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Timer";
