+++ shared/components/UI/HeaderButton/index.js
@@ -1,17 +1,1 @@
-import React from "react";
-import { Platform } from "react-native";
-import { HeaderButton } from "react-navigation-header-buttons";
-import { Ionicons } from "@expo/vector-icons";
-
-const CustomHeaderButton = (props) => {
-    return (
-        <HeaderButton
-            {...props}
-            IconComponent={Ionicons}
-            iconSize={23}
-            color={Platform.OS === "android" ? "white" : props.color}
-        />
-    );
-};
-
-export default CustomHeaderButton;
+export { default } from "./HeaderButton";
