+++ shared/constants/Colors/Colors.js
@@ -1,0 +1,3 @@
+export const Success = "#53B561";
+export const Notice = "#e39768";
+export const Caution = "#e36868";
