+++ shared/constants/Colors/index.js
@@ -1,3 +1,1 @@
-export const Success = "#53B561";
-export const Notice = "#e39768";
-export const Caution = "#e36868";
+export { Success, Notice, Caution } from "./Colors";
