+++ shared/models/task/index.js
@@ -1,9 +1,1 @@
-class Task {
-    constructor(id, title, count = 0) {
-        this.id = id;
-        this.title = title;
-        this.count = count;
-    }
-}
-
-export default Task;
+export { default } from "./task";
