+++ shared/models/task/task.js
@@ -1,0 +1,9 @@
+class Task {
+    constructor(id, title, count = 0) {
+        this.id = id;
+        this.title = title;
+        this.count = count;
+    }
+}
+
+export default Task;
