+++ shared/navigation/StartUpNavigator/StartUpNavigator.js
