+++ shared/navigation/StartUpNavigator/index.js
@@ -1,0 +1,1 @@
+export { StartUpNavigator } from "./StartUpNavigator";
